完成文件资源上传功能

+++ cloud-cms/cloud-cms-web/src/main/java/com/lyl/study/cloud/cms/web/controller/ResourceEntityController.java
@@ -26,7 +26,8 @@ import java.nio.file.Paths;
 
 import static com.lyl.study.cloud.cms.api.CmsErrorCode.*;
 
-@RestController("/resourceEntity")
+@RestController
+@RequestMapping("/resourceEntity")
 public class ResourceEntityController implements InitializingBean {
     protected final Logger logger = LoggerFactory.getLogger(this.getClass());
 
@@ -46,24 +47,79 @@ public class ResourceEntityController implements InitializingBean {
     @Reference
     private ResourceEntityFacade resourceEntityFacade;
 
+    public Long getMaxImageSize() {
+        return maxImageSize;
+    }
+
+    public void setMaxImageSize(Long maxImageSize) {
+        this.maxImageSize = maxImageSize;
+    }
+
+    public Long getMaxAudioSize() {
+        return maxAudioSize;
+    }
+
+    public void setMaxAudioSize(Long maxAudioSize) {
+        this.maxAudioSize = maxAudioSize;
+    }
+
+    public Long getMaxVideoSize() {
+        return maxVideoSize;
+    }
+
+    public void setMaxVideoSize(Long maxVideoSize) {
+        this.maxVideoSize = maxVideoSize;
+    }
+
+    public Long getMaxFileSize() {
+        return maxFileSize;
+    }
+
+    public void setMaxFileSize(Long maxFileSize) {
+        this.maxFileSize = maxFileSize;
+    }
+
+    public String getUrlPrefix() {
+        return urlPrefix;
+    }
+
+    public void setUrlPrefix(String urlPrefix) {
+        this.urlPrefix = urlPrefix;
+    }
+
+    public Path getUploadPath() {
+        return uploadPath;
+    }
+
+    public void setUploadPath(Path uploadPath) {
+        this.uploadPath = uploadPath;
+    }
+
     @Override
     public void afterPropertiesSet() {
+        logger.info("上传文件配置: {}", properties);
         BeanUtils.copyProperties(properties, this);
     }
 
     @PostMapping
     public Result uploadFile(@RequestParam("file") MultipartFile multipartFile) {
         try {
+            // 校验文件大小
             checkFileSize(multipartFile);
+            // 计算文件保存地址
             Path path = resolveFilePath(multipartFile);
+            // 保存到本地
+            logger.info("保存文件: {}", path.toAbsolutePath().toString());
             multipartFile.transferTo(path.toAbsolutePath().toFile());
+            // 创建数据库记录保存表单
             ResourceEntitySaveForm form = buildSaveForm(multipartFile, path);
 
             try {
                 resourceEntityFacade.save(form);
                 return new Result<>(OK, "上传成功", form.getUrl());
             } catch (Exception e) {
-                logger.error(e.toString());
+                logger.error("内部服务调用错误，将回滚上传的文件: {}\n\n{}",
+                        path.toAbsolutePath().toString(), e.toString());
                 Files.deleteIfExists(Paths.get(form.getFilepath()));
                 return new Result<>(FILE_SAVE_INTERNAL_SERVICE_EXCEPTION, "内部服务调用错误", e.getMessage());
             }
@@ -74,8 +130,20 @@ public class ResourceEntityController implements InitializingBean {
 
     @DeleteMapping("/{id}")
     public Result deleteById(@PathVariable("id") Long id) {
-        int rows = resourceEntityFacade.deleteById(id);
-        if (rows > 0) {
+        logger.info("删除ID为" + id + "的资源文件");
+
+        ResourceEntityDTO dto = resourceEntityFacade.getById(id);
+
+        if (dto != null) {
+            try {
+                Files.delete(Paths.get(dto.getFilepath()));
+                logger.info("文件删除成功: {}", dto.getFilepath());
+            } catch (IOException e) {
+                // 可能是文件或路径不正确，暂不处理
+                logger.error("文件删除失败: {}", e.getMessage());
+            }
+
+            resourceEntityFacade.deleteById(id);
             return new Result<>(OK, "删除成功", null);
         } else {
             return new Result<>(NOT_FOUND, "找不到ID为" + id + "的资源文件", null);
@@ -85,6 +153,7 @@ public class ResourceEntityController implements InitializingBean {
     @GetMapping("/list")
     public Result<PageInfo<ResourceEntityDTO>> list(@ModelAttribute ResourceEntityListConditions conditions) {
         PageInfo<ResourceEntityDTO> page = resourceEntityFacade.list(conditions);
+        page.getRecords().forEach(entity -> entity.setFilepath(null));
         return new Result<>(OK, "查询成功", page);
     }
 
@@ -137,14 +206,14 @@ public class ResourceEntityController implements InitializingBean {
     private void checkFileSize(MultipartFile file) {
         MediaType mediaType = MediaType.parseMediaType(file.getContentType());
 
-        if (MEDIA_TYPE_IMAGE.includes(mediaType)) {
-            Assert.isTrue(maxImageSize <= file.getSize(), "上传图片文件过大");
-        } else if (MEDIA_TYPE_AUDIO.includes(mediaType)) {
-            Assert.isTrue(maxAudioSize <= file.getSize(), "上传音频文件过大");
-        } else if (MEDIA_TYPE_VIDEO.includes(mediaType)) {
-            Assert.isTrue(maxVideoSize <= file.getSize(), "上传视频文件过大");
+        if (MEDIA_TYPE_IMAGE.includes(mediaType) && maxImageSize != null) {
+            Assert.isTrue(maxImageSize > file.getSize(), "上传图片文件过大");
+        } else if (MEDIA_TYPE_AUDIO.includes(mediaType) && maxAudioSize != null) {
+            Assert.isTrue(maxAudioSize > file.getSize(), "上传音频文件过大");
+        } else if (MEDIA_TYPE_VIDEO.includes(mediaType) && maxVideoSize != null) {
+            Assert.isTrue(maxVideoSize > file.getSize(), "上传视频文件过大");
         } else {
-            Assert.isTrue(maxFileSize <= file.getSize(), "上传文件过大");
+            Assert.isTrue(maxFileSize > file.getSize(), "上传文件过大");
         }
     }
 }
